rename index files to their component names
