+++ webview-ui/src/components/mcp/configuration/tabs/installed/InstalledServersView.tsx
@@ -1,7 +1,7 @@
 import { VSCodeButton, VSCodeLink } from "@vscode/webview-ui-toolkit/react"
 import { vscode } from "../../../../../utils/vscode"
 import { useExtensionState } from "../../../../../context/ExtensionStateContext"
-import ServerRow from "./server-row"
+import ServerRow from "./server-row/ServerRow"
 const InstalledServersView = () => {
 	const { mcpServers: servers } = useExtensionState()
